T-6951 local/rb_sources/rb_source_dp_course.php: Changing report base
This is the source for the record of learning (plan) course embedded report.
Previously it showed all courses for which the user had started course
completion. We've changed it so that instead it shows all courses in which
the user has been assigned a role, and/or the course is in at least
one of their plans.