MDL-24561 forum subscription checks sesskey - backported from 2.0
The only case when sesskey is not available is when we came to
subscribe.php from a link in email. In that case, we display a
confirmation page that redirects back providing the sesskey.
All other links to subscribe.php are fixed now so they provide sesskey
as a parameter.

Change-Id: I6fc08ec7f588862dfd4ba44d80c740c9e2b56537
Reviewed-on: http://review.dev.totaralms.com/115
Reviewed-by: Aaron Barnes <aaronb@catalyst.net.nz>
Tested-by: Aaron Barnes <aaronb@catalyst.net.nz>