T-7186 hierarchy: Removed custom field column from assigned competencies
Because custom fields are assigned to depth levels, it doesn't make sense to show custom fields as columns
in a table when the rows can contain competencies from different depth levels. The current code didn't
take this into account and was broken so I removed the custom field columns entirely.